Change style

+++ src/main/java/com/fiap/enrico_andrade/security/repository/UserRepository.java
@@ -4,6 +4,6 @@ import com.fiap.enrico_andrade.security.entity.AppUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<AppUser, Long> {
+public interface UserRepository extends JpaRepository<AppUser, Integer> {
     Optional<AppUser> findByUsername(String username);
 }
